+++ apm-agent-plugins/apm-spring-resttemplate-plugin/src/main/java/co/elastic/apm/agent/resttemplate/SpringRestTemplateInstrumentation.java
@@ -37,7 +37,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Objects;
 
-import static co.elastic.apm.agent.http.client.HttpClientHelper.HTTP_CLIENT_SPAN_TYPE_PREFIX;
 import static net.bytebuddy.matcher.ElementMatchers.declaresMethod;
 import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
 import static net.bytebuddy.matcher.ElementMatchers.isInterface;
@@ -49,8 +48,6 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 public class SpringRestTemplateInstrumentation extends ElasticApmInstrumentation {
 
-    private static final String SPAN_TYPE_SPRING_REST_TEMPLATE = HTTP_CLIENT_SPAN_TYPE_PREFIX + "spring-resttemplate";
-
     @Advice.OnMethodEnter(suppress = Throwable.class)
     private static void beforeExecute(@Advice.This ClientHttpRequest request,
                                       @Advice.Local("span") Span span) {
@@ -59,7 +56,7 @@ public class SpringRestTemplateInstrumentation extends ElasticApmInstrumentation
         }
         final TraceContextHolder<?> parent = tracer.getActive();
         span = HttpClientHelper.startHttpClientSpan(parent, Objects.toString(request.getMethod()), request.getURI(),
-            request.getURI().getHost(), SPAN_TYPE_SPRING_REST_TEMPLATE);
+            request.getURI().getHost());
         if (span != null) {
             request.getHeaders().add(TraceContext.TRACE_PARENT_HEADER, span.getTraceContext().getOutgoingTraceParentHeader().toString());
         }